add re-usable BratBuilder within builders package; move Brat-related types into that
