+++ src/pie_datasets/builders/__init__.py
@@ -1,0 +1,1 @@
+from .brat import BratBuilder, BratConfig
