+++ tests/dataset_builders/pie/test_brat.py
@@ -6,15 +6,13 @@ from pytorch_ie.annotations import BinaryRelation, LabeledMultiSpan, LabeledSpan
 from pytorch_ie.core import Annotation
 from pytorch_ie.documents import TextBasedDocument
 
-from dataset_builders.pie.brat.brat import (
-    BratDatasetLoader,
-    document_to_example,
-    example_to_document,
-)
-from pie_datasets.document.types import (
-    Attribute,
+from dataset_builders.pie.brat.brat import BratBuilder
+from pie_datasets.builders.brat import (
+    BratAttribute,
     BratDocument,
     BratDocumentWithMergedSpans,
+    document_to_example,
+    example_to_document,
 )
 from tests.dataset_builders.common import PIE_BASE_PATH, PIE_DS_FIXTURE_DATA_PATH
 
@@ -22,7 +20,7 @@ datasets.disable_caching()
 
 DATASET_NAME = "brat"
 PIE_DATASET_PATH = PIE_BASE_PATH / DATASET_NAME
-HF_DATASET_PATH = BratDatasetLoader.BASE_DATASET_PATH
+HF_DATASET_PATH = BratBuilder.BASE_DATASET_PATH
 FIXTURE_DATA_PATH = PIE_DS_FIXTURE_DATA_PATH / DATASET_NAME
 SPLIT_SIZES = {"train": 2}
 
@@ -43,7 +41,7 @@ def resolve_annotation(annotation: Annotation) -> Any:
             annotation.label,
             resolve_annotation(annotation.tail),
         )
-    elif isinstance(annotation, Attribute):
+    elif isinstance(annotation, BratAttribute):
         result = (resolve_annotation(annotation.annotation), annotation.label)
         if annotation.value is not None:
             return result + (annotation.value,)
@@ -135,7 +133,7 @@ def test_hf_example(hf_example, sample_idx):
 
 
 @pytest.fixture(
-    params=[config.name for config in BratDatasetLoader.BUILDER_CONFIGS],  # scope="module"
+    params=[config.name for config in BratBuilder.BUILDER_CONFIGS],  # scope="module"
 )
 def pie_dataset_variant(request):
     return request.param
@@ -145,7 +143,7 @@ def pie_dataset_variant(request):
 def generated_document(
     hf_example, hf_dataset, pie_dataset_variant
 ) -> Union[BratDocument, BratDocumentWithMergedSpans]:
-    builder = BratDatasetLoader(name=pie_dataset_variant)
+    builder = BratBuilder(name=pie_dataset_variant)
     kwargs = builder._generate_document_kwargs(hf_dataset["train"]) or {}
     document = builder._generate_document(example=hf_example, **kwargs)
     assert document is not None
